+++ src/main/java/ua/com/foxminded/studenthostel/models/mappers/FloorMapper.java
@@ -4,6 +4,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 import ua.com.foxminded.studenthostel.models.Floor;
 
+import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -13,7 +14,8 @@ public class FloorMapper implements RowMapper<Floor> {
     @Override
     public Floor mapRow(ResultSet resultSet, int i) throws SQLException {
         Floor floor = new Floor();
-        floor.setId(resultSet.getInt("floor_id"));
+
+        floor.setId(BigInteger.valueOf(resultSet.getInt("floor_id")));
         floor.setName(resultSet.getString("floor_name"));
 
         return floor;
