+++ src/main/java/ua/com/foxminded/studenthostel/models/mappers/GroupMapper.java
@@ -5,6 +5,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 import ua.com.foxminded.studenthostel.models.Group;
 
+import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -14,10 +15,11 @@ public class GroupMapper implements RowMapper<Group> {
     @Override
     public Group mapRow(ResultSet resultSet, int i) throws SQLException {
         Group group = new Group();
+
         group.setName(resultSet.getString("group_name"));
-        group.setId(resultSet.getInt("group_id"));
-        group.setCourseNumberId(resultSet.getInt("course_number_id"));
-        group.setFacultyId(resultSet.getInt("faculty_id"));
+        group.setId(BigInteger.valueOf(resultSet.getInt("group_id")));
+        group.setCourseNumberId(BigInteger.valueOf(resultSet.getInt("course_number_id")));
+        group.setFacultyId(BigInteger.valueOf(resultSet.getInt("faculty_id")));
 
         return group;
     }
