+++ src/main/java/ua/com/foxminded/studenthostel/models/mappers/StudentMapper.java
@@ -4,6 +4,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 import ua.com.foxminded.studenthostel.models.Student;
 
+import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
@@ -13,12 +14,13 @@ public class StudentMapper implements RowMapper<Student> {
     public Student mapRow(ResultSet resultSet, int i) throws SQLException {
 
         Student student = new Student();
-        student.setId(resultSet.getInt("student_id"));
+
+        student.setId(BigInteger.valueOf(resultSet.getInt("student_id")));
         student.setFirstName(resultSet.getString("first_name"));
         student.setLastName(resultSet.getString("last_name"));
         student.setHoursDebt(resultSet.getInt("hours_debt"));
-        student.setGroupId(resultSet.getInt("group_id"));
-        student.setRoomId(resultSet.getInt("room_id"));
+        student.setGroupId(BigInteger.valueOf(resultSet.getInt("group_id")));
+        student.setRoomId(BigInteger.valueOf(resultSet.getInt("room_id")));
 
         return student;
     }
